requirements.txt
cffi==1.14.0
Django==2.2.5
django-adminlte2==0.4.1
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.11.0
misaka==2.1.0
mysqlclient==1.4.3
Pillow==5.4.1
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycparser==2.20
pyodbc==4.0.30
pytz==2019.3
reportlab==3.5.13
sqlparse==0.3.1
virtualenv==16.7.4
